+++ packages/form-upload/tests/UploadProgressBar.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, cleanup, waitFor } from '@testing-library/react';
+import { render, cleanup, waitFor, fireEvent } from '@testing-library/react';
 import { act } from 'react-dom/test-utils';
 import MockUpload from './mockUpload';
 
@@ -61,6 +61,39 @@ describe('UploadProgressBar', () => {
 
     const el = getByTestId('password-form-encrypted');
     expect(el).toBeDefined();
+  });
+
+  test('should render password modal', () => {
+    const { getByTestId, getByText } = render(<UploadProgressBar upload={instance} />);
+    instance.error('Encrypted files require a password', 'encrypted');
+
+    const enterPassword = getByText('Enter password');
+    fireEvent.click(enterPassword);
+
+    const modal = getByTestId('password-form-modal');
+    expect(modal).toBeDefined();
+  });
+
+  test('should not submit parent forms on password submit', () => {
+    const submitFunction = jest.fn();
+
+    const { getByTestId, getByText } = render(
+      <form onSubmit={submitFunction}>
+        <UploadProgressBar upload={instance} />
+      </form>
+    );
+    instance.error('Encrypted files require a password', 'encrypted');
+
+    const enterPassword = getByText('Enter password');
+    fireEvent.click(enterPassword);
+
+    const modal = getByTestId('password-form-modal');
+    expect(modal).toBeDefined();
+
+    const submit = getByText('Ok');
+    fireEvent.click(submit);
+
+    expect(submitFunction).not.toHaveBeenCalled();
   });
 
   test('should render striped', () => {