+++ packages/page-header/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 11.2.6 (2022-05-20)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
+## [11.2.5](https://github.com/Availity/availity-react/compare/@availity/page-header@11.2.4...@availity/page-header@11.2.5) (2022-05-18)
+
+**Note:** Version bump only for package @availity/page-header
+
+
+
+
+
 ## [11.2.4](https://github.com/Availity/availity-react/compare/@availity/page-header@11.2.3...@availity/page-header@11.2.4) (2022-05-11)
 
 **Note:** Version bump only for package @availity/page-header